// app_asistentes/src/app/interfaces/users.ts
export interface Users {
    id: number;
    username: string;
    email: string;
    password: string;
    isactive: boolean;
    name: string;
    age: number;
    rut: string;
    dob: string;
    gender: string;
    profilePicture: string;
}

export interface User {
    username: string;
    email: string;
    password: string;
    isactive: boolean;
    name: string;
    age: number;
    rut: string;
    dob: string;
    gender: string;
    profilePicture: string;
}